Move appConfig into config/app
- Merged variables in appConfig into config/app
- Removed appConfig as it was confusing
