+++ config/app.js
@@ -1,8 +1,12 @@
+import { DB_TYPES } from './dbTypes';
+
 export const HOST = process.env.HOSTNAME || 'localhost';
 export const PORT = process.env.PORT || '3000';
+export const DB_TYPE = process.env.DB_TYPE || DB_TYPES.MONGO;
+export const ENV = process.env.NODE_ENV || 'development';
 
-export const isProduction = process.env.NODE_ENV === 'production';
-export const isDebug = process.env.NODE_ENV === 'development';
+export const isProduction = ENV === 'production';
+export const isDebug = ENV === 'development';
 export const isClient = typeof window !== 'undefined';
 
 export const baseURL = `http://${HOST}:${PORT}`;
