+++ docs/databases.md
@@ -5,7 +5,7 @@
 Change the following:
 ```javascript
 /*
- * server/config/appConfig.js
+ * config/app.js
  */
 
 DB_TYPE: process.env.DB_TYPE || DB_TYPES.NONE
@@ -19,13 +19,12 @@ Currently, we support these DB_TYPES:
 - POSTGRES
 - NONE
 
-We abstracted the DB config in `appConfig` to enable you to require the correct files if you were to use a different database, e.g. postgresql.
+We abstracted the DB config in `app` to enable you to require the correct files if you were to use a different database, e.g. postgresql.
 
-Please read [this PR](https://github.com/choonkending/react-webpack-node/pull/190#issuecomment-210273745) for more context.
 
 ```javascript
 /*
- * server/config/appConfig.js
+ * config/app.js
  */
 
 DB_TYPE: process.env.DB_TYPE || DB_TYPES.YOUR_DB
