+++ server/config/express.js
@@ -9,7 +9,7 @@ import gzip from 'compression';
 import helmet from 'helmet';
 import unsupportedMessage from '../db/unsupportedMessage';
 import { sessionSecret } from './secrets';
-import { DB_TYPE, ENV } from './appConfig';
+import { DB_TYPE, ENV } from '../../config/app';
 import { session as dbSession } from '../db';
 
 export default (app) => {
