+++ server/db/sequelize/models/index.js
@@ -1,6 +1,6 @@
 import Sequelize from 'sequelize';
 import sequelizeConfig from '../sequelize_config';
-import { ENV } from '../../../config/appConfig';
+import { ENV } from '../../../../config/app';
 import tokenModel from './tokens';
 import topicModel from './topics';
 import userModel from './users';
