+++ server/db/sequelize/sequelize_config.js
@@ -1,4 +1,4 @@
-import { DB_TYPE } from '../../config/appConfig';
+import { DB_TYPE } from '../../../config/app';
 import { DB_TYPES } from '../../../config/dbTypes';
 
 switch (DB_TYPE) {
